backend_py/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
sse-starlette==2.1.3
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
prometheus-client==0.21.1
redis==5.2.1
requests==2.32.3
httpx==0.28.1
